verzamelen/01-level/Van Archief naar Database/requirements.txt
mysql-connector==2.2.9
mysql-connector-python==8.0.28
protobuf==3.20.1
